requirements.txt
locust==2.32.0
pyyaml==6.0.2